[DX] Extract batch events
