+++ src/Akeneo/Bundle/BatchBundle/EventListener/LoggerSubscriber.php
@@ -3,9 +3,9 @@
 namespace Akeneo\Bundle\BatchBundle\EventListener;
 
 use Akeneo\Component\Batch\Event\EventInterface;
-use Akeneo\Bundle\BatchBundle\Event\InvalidItemEvent;
-use Akeneo\Bundle\BatchBundle\Event\JobExecutionEvent;
-use Akeneo\Bundle\BatchBundle\Event\StepExecutionEvent;
+use Akeneo\Component\Batch\Event\InvalidItemEvent;
+use Akeneo\Component\Batch\Event\JobExecutionEvent;
+use Akeneo\Component\Batch\Event\StepExecutionEvent;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Translation\TranslatorInterface;
@@ -231,7 +231,7 @@ class LoggerSubscriber implements EventSubscriberInterface
     /**
      * Log invalid item event
      *
-     * @param InvalidItemEvent $event
+     * @param \Akeneo\Component\Batch\Event\InvalidItemEvent $event
      */
     public function invalidItem(InvalidItemEvent $event)
     {
