+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/EventListener/NotificationSubscriberTest.php
@@ -70,7 +70,7 @@ class NotificationSubscriberTest extends \PHPUnit_Framework_TestCase
     private function getJobExecutionEventMock($jobExecution = null)
     {
         $event = $this
-            ->getMockBuilder('Akeneo\Bundle\BatchBundle\Event\JobExecutionEvent')
+            ->getMockBuilder('Akeneo\Component\Batch\Event\JobExecutionEvent')
             ->disableOriginalConstructor()
             ->getMock();
