+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/EventListener/SetJobExecutionLogFileSubscriberTest.php
@@ -61,7 +61,7 @@ class SetJobExecutionLogFileSubscriberTest extends \PHPUnit_Framework_TestCase
     private function getJobExecutionEventMock($jobExecution = null)
     {
         $event = $this
-            ->getMockBuilder('Akeneo\Bundle\BatchBundle\Event\JobExecutionEvent')
+            ->getMockBuilder('Akeneo\Component\Batch\Event\JobExecutionEvent')
             ->disableOriginalConstructor()
             ->getMock();
