+++ src/Pim/Bundle/BaseConnectorBundle/EventListener/InvalidItemsCollector.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\BaseConnectorBundle\EventListener;
 
 use Akeneo\Component\Batch\Event\EventInterface;
-use Akeneo\Bundle\BatchBundle\Event\InvalidItemEvent;
+use Akeneo\Component\Batch\Event\InvalidItemEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -31,7 +31,7 @@ class InvalidItemsCollector implements EventSubscriberInterface
     /**
      * Collect unique invalid items
      *
-     * @param InvalidItemEvent $event
+     * @param \Akeneo\Component\Batch\Event\InvalidItemEvent $event
      */
     public function collect(InvalidItemEvent $event)
     {
